Update class-list-table.php
To help chase down PHP errors, I expanded the search capability in "search_callback()" to allow for narrowing search across snippets by line number in the code field only.  This is accomplished by parsing the search term for "@#:" and splitting the search term into three parts: <search term>@#:<line number>. Searches without "@#:" are handled normally.

NB: Since the search term is converted to entities when submitted, it creates a re-display problem of truncating the search term when the results are returned. I just convert $_REQUEST['s'] back to characters and didn't consider UTF-8. I'm not sure if this is a problem or not?

Hope you find this useful.

Mike